Update pinellas transform scripts
