+++ pinellas/scripts/data_extractor.js
@@ -40,7 +40,7 @@ const extraFeaturesDescriptionMappings = [
 function extractExtraFeatures($, dataDir, requestIdentifier, sourceHttpRequest) {
   const rows = $("#tblExtraFeatures tbody tr");
   if (!rows || rows.length === 0) {
-    ["propertyLot.json", "propertyUtility.json", "propertyStructure.json", "utility.json"].forEach((filename) => {
+    ["propertyLot.json", "propertyUtility.json", "propertyStructure.json"].forEach((filename) => {
       const filePath = path.join(dataDir, filename);
       if (fs.existsSync(filePath)) {
         fs.unlinkSync(filePath);
@@ -69,57 +69,13 @@ function extractExtraFeatures($, dataDir, requestIdentifier, sourceHttpRequest) 
     }
   });
 
-  // CRITICAL: Only lot and structure files should be created here
-  // Utility files are created as utility_N.json with proper relationships later (lines 3431-3490)
-  // NEVER add "utility" to this fileMap - it will cause an unused file error
   const fileMap = {
     lot: "lot.json",
+    utility: "utility.json",
     structure: "structure.json",
-    // utility: "utility.json", // NEVER uncomment this - use utility_N.json files instead
   };
 
-  // Clean up utility.json if it exists (no longer used - we use utility_N.json instead)
-  // Note: utility data is collected in grouped.utility but NOT written as a single file
-  // Instead, utilities are written as utility_N.json with proper relationships later in the script
-  const utilityFilePath = path.join(dataDir, "utility.json");
-  if (fs.existsSync(utilityFilePath)) {
-    try {
-      fs.unlinkSync(utilityFilePath);
-      console.log("Removed utility.json during extractExtraFeatures - using utility_N.json files instead");
-    } catch (e) {
-      console.error("Failed to remove utility.json:", e);
-    }
-  }
-
-  // Additional safeguard: Delete utility.json before any file writing to prevent accidental generation
-  const utilityJsonBeforeWritePath = path.join(dataDir, "utility.json");
-  if (fs.existsSync(utilityJsonBeforeWritePath)) {
-    try {
-      fs.unlinkSync(utilityJsonBeforeWritePath);
-      console.log("Pre-write safeguard: Removed utility.json before file writing");
-    } catch (e) {
-      console.error("Failed to remove utility.json in pre-write safeguard:", e);
-    }
-  }
-
-  // FINAL safeguard before the loop: Force delete utility.json to prevent any accidental creation
-  try {
-    const finalUtilityCheck = path.join(dataDir, "utility.json");
-    if (fs.existsSync(finalUtilityCheck)) {
-      fs.unlinkSync(finalUtilityCheck);
-      console.log("Final pre-loop cleanup: Removed utility.json before fileMap processing");
-    }
-  } catch (e) {
-    console.error("Failed to remove utility.json in final pre-loop cleanup:", e);
-  }
-
   Object.entries(fileMap).forEach(([cls, filename]) => {
-    // Safety check: Never write utility.json (we use utility_N.json instead)
-    if (filename === "utility.json" || cls === "utility") {
-      console.log(`Skipping ${filename} - using utility_N.json files with relationships instead`);
-      return;
-    }
-
     const props = grouped[cls] || {};
     const hasProperties = Object.keys(props).length > 0;
     const filePath = path.join(dataDir, filename);
@@ -137,17 +93,6 @@ function extractExtraFeatures($, dataDir, requestIdentifier, sourceHttpRequest) 
     };
     writeJSON(filePath, output);
   });
-
-  // IMMEDIATE post-loop cleanup: Delete utility.json if it was somehow created
-  try {
-    const postLoopUtilityCheck = path.join(dataDir, "utility.json");
-    if (fs.existsSync(postLoopUtilityCheck)) {
-      fs.unlinkSync(postLoopUtilityCheck);
-      console.log("Post-loop cleanup: Removed utility.json after fileMap processing");
-    }
-  } catch (e) {
-    console.error("Failed to remove utility.json in post-loop cleanup:", e);
-  }
 }
 
 const propertyTypeMapping=[
@@ -217,14 +162,6 @@ const propertyTypeMapping=[
   },
   {
     "property_usecode": "0110 SINGLE FAMILY HOME",
-    "ownership_estate_type": "FeeSimple",
-    "build_status": "Improved",
-    "structure_form": "SingleFamilyDetached",
-    "property_usage_type": "Residential",
-    "property_type": "Building"
-  },
-  {
-    "property_usecode": "0111 SINGLE FAMILY COMMUNITY LAND TRUST",
     "ownership_estate_type": "FeeSimple",
     "build_status": "Improved",
     "structure_form": "SingleFamilyDetached",
@@ -313,14 +250,6 @@ const propertyTypeMapping=[
   },
   {
     "property_usecode": "0436 CONDO CONVERSION - APARTMENTS TO PLATTED CONDO (PREDOMINATELY OWNER-OCCUPIED)",
-    "ownership_estate_type": "Condominium",
-    "build_status": "Improved",
-    "structure_form": "ApartmentUnit",
-    "property_usage_type": "Residential",
-    "property_type": "Unit"
-  },
-  {
-    "property_usecode": "0437 CONDO COM APARTMENTS",
     "ownership_estate_type": "Condominium",
     "build_status": "Improved",
     "structure_form": "ApartmentUnit",
@@ -1607,19 +1536,9 @@ const propertyUsageTypeByUseCode = propertyTypeMapping.reduce((lookup, entry) =>
 }, {});
 function mapPropertyTypeFromUseCode(code) {
   if (!code && code !== 0) return null;
-  const codeStr = String(code).trim();
-  
-  // First try matching by 4-digit code prefix
-  const codeMatch = codeStr.match(/^(\d{4})/);
-  if (codeMatch) {
-    const codePrefix = codeMatch[1];
-    const entry = propertyTypeMapping.find(item => item.property_usecode.startsWith(codePrefix));
-    if (entry) return entry.property_type;
-  }
-  
-  // Fallback to full normalization matching
-  const normalizedInput = codeStr.replace(/\s+/g, "").toUpperCase();
-  if (normalizedInput && Object.prototype.hasOwnProperty.call(propertyTypeByUseCode, normalizedInput)) {
+  const normalizedInput = String(code).replace(/[\s:]+/g, "").toUpperCase();
+  if (!normalizedInput) return null;
+  if (Object.prototype.hasOwnProperty.call(propertyTypeByUseCode, normalizedInput)) {
     return propertyTypeByUseCode[normalizedInput];
   }
   return null;
@@ -1627,19 +1546,9 @@ function mapPropertyTypeFromUseCode(code) {
 
 function mapOwnershipEstateTypeFromUseCode(code) {
   if (!code && code !== 0) return null;
-  const codeStr = String(code).trim();
-  
-  // First try matching by 4-digit code prefix
-  const codeMatch = codeStr.match(/^(\d{4})/);
-  if (codeMatch) {
-    const codePrefix = codeMatch[1];
-    const entry = propertyTypeMapping.find(item => item.property_usecode.startsWith(codePrefix));
-    if (entry) return entry.ownership_estate_type;
-  }
-  
-  // Fallback to full normalization matching
-  const normalizedInput = codeStr.replace(/\s+/g, "").toUpperCase();
-  if (normalizedInput && Object.prototype.hasOwnProperty.call(ownershipEstateTypeByUseCode, normalizedInput)) {
+  const normalizedInput = String(code).replace(/[\s:]+/g, "").toUpperCase();
+  if (!normalizedInput) return null;
+  if (Object.prototype.hasOwnProperty.call(ownershipEstateTypeByUseCode, normalizedInput)) {
     return ownershipEstateTypeByUseCode[normalizedInput];
   }
   return null;
@@ -1647,19 +1556,9 @@ function mapOwnershipEstateTypeFromUseCode(code) {
 
 function mapBuildStatusFromUseCode(code) {
   if (!code && code !== 0) return null;
-  const codeStr = String(code).trim();
-  
-  // First try matching by 4-digit code prefix
-  const codeMatch = codeStr.match(/^(\d{4})/);
-  if (codeMatch) {
-    const codePrefix = codeMatch[1];
-    const entry = propertyTypeMapping.find(item => item.property_usecode.startsWith(codePrefix));
-    if (entry) return entry.build_status;
-  }
-  
-  // Fallback to full normalization matching
-  const normalizedInput = codeStr.replace(/\s+/g, "").toUpperCase();
-  if (normalizedInput && Object.prototype.hasOwnProperty.call(buildStatusByUseCode, normalizedInput)) {
+  const normalizedInput = String(code).replace(/[\s:]+/g, "").toUpperCase();
+  if (!normalizedInput) return null;
+  if (Object.prototype.hasOwnProperty.call(buildStatusByUseCode, normalizedInput)) {
     return buildStatusByUseCode[normalizedInput];
   }
   return null;
@@ -1667,19 +1566,9 @@ function mapBuildStatusFromUseCode(code) {
 
 function mapStructureFormFromUseCode(code) {
   if (!code && code !== 0) return null;
-  const codeStr = String(code).trim();
-  
-  // First try matching by 4-digit code prefix
-  const codeMatch = codeStr.match(/^(\d{4})/);
-  if (codeMatch) {
-    const codePrefix = codeMatch[1];
-    const entry = propertyTypeMapping.find(item => item.property_usecode.startsWith(codePrefix));
-    if (entry) return entry.structure_form;
-  }
-  
-  // Fallback to full normalization matching
-  const normalizedInput = codeStr.replace(/\s+/g, "").toUpperCase();
-  if (normalizedInput && Object.prototype.hasOwnProperty.call(structureFormByUseCode, normalizedInput)) {
+  const normalizedInput = String(code).replace(/[\s:]+/g, "").toUpperCase();
+  if (!normalizedInput) return null;
+  if (Object.prototype.hasOwnProperty.call(structureFormByUseCode, normalizedInput)) {
     return structureFormByUseCode[normalizedInput];
   }
   return null;
@@ -1687,19 +1576,9 @@ function mapStructureFormFromUseCode(code) {
 
 function mapPropertyUsageTypeFromUseCode(code) {
   if (!code && code !== 0) return null;
-  const codeStr = String(code).trim();
-  
-  // First try matching by 4-digit code prefix
-  const codeMatch = codeStr.match(/^(\d{4})/);
-  if (codeMatch) {
-    const codePrefix = codeMatch[1];
-    const entry = propertyTypeMapping.find(item => item.property_usecode.startsWith(codePrefix));
-    if (entry) return entry.property_usage_type;
-  }
-  
-  // Fallback to full normalization matching
-  const normalizedInput = codeStr.replace(/\s+/g, "").toUpperCase();
-  if (normalizedInput && Object.prototype.hasOwnProperty.call(propertyUsageTypeByUseCode, normalizedInput)) {
+  const normalizedInput = String(code).replace(/[\s:]+/g, "").toUpperCase();
+  if (!normalizedInput) return null;
+  if (Object.prototype.hasOwnProperty.call(propertyUsageTypeByUseCode, normalizedInput)) {
     return propertyUsageTypeByUseCode[normalizedInput];
   }
   return null;
@@ -2309,14 +2188,6 @@ function extractSalesHistoryFiles($, dataDir, requestIdentifier, sourceHttpReque
       ipfs_url: null,
     };
     writeJSON(path.join(dataDir, fileName), payload);
-
-    // Create property_has_file relationship
-    const relationshipFileName = `relationship_property_has_file_${idx + 1}.json`;
-    const relationship = {
-      from: { "/": "./property.json" },
-      to: { "/": `./${fileName}` }
-    };
-    writeJSON(path.join(dataDir, relationshipFileName), relationship);
   });
 }
 
@@ -2843,44 +2714,6 @@ function extract() {
   const dataDir = path.join("data");
   ensureDir(dataDir);
 
-  // CRITICAL FIRST CHECK: Before anything else, ensure utility.json doesn't exist from previous runs
-  // This is the first line of defense against the "Unused data JSON file" error
-  const earlyUtilityCheck = path.join(dataDir, "utility.json");
-  if (fs.existsSync(earlyUtilityCheck)) {
-    try {
-      fs.unlinkSync(earlyUtilityCheck);
-      console.log("PRE-EXECUTION: Removed utility.json from previous run");
-    } catch (e) {
-      console.error("Failed to remove utility.json during pre-execution cleanup:", e);
-    }
-  }
-
-  // CRITICAL: utility.json should NEVER be created - we use utility_N.json files with proper relationships
-  // This cleanup ensures any legacy utility.json from previous executions is removed
-  const utilityJsonPath = path.join(dataDir, "utility.json");
-  if (fs.existsSync(utilityJsonPath)) {
-    try {
-      fs.unlinkSync(utilityJsonPath);
-      console.log("Initial cleanup: Removed utility.json - using utility_N.json files with relationships instead");
-    } catch (e) {
-      console.error("Failed to remove utility.json during initial cleanup:", e);
-    }
-  }
-
-  // Additional cleanup for other legacy utility file variants
-  const legacyUtilityFiles = ["propertyUtility.json", "property_utility.json"];
-  legacyUtilityFiles.forEach(filename => {
-    const filepath = path.join(dataDir, filename);
-    if (fs.existsSync(filepath)) {
-      try {
-        fs.unlinkSync(filepath);
-        console.log(`Initial cleanup: Removed legacy file ${filename}`);
-      } catch (e) {
-        console.error(`Failed to remove ${filename}:`, e);
-      }
-    }
-  });
-
   const html = fs.readFileSync("input.html", "utf8");
   const $ = cheerio.load(html);
   const unAddr = readJSON("unnormalized_address.json") || {};
@@ -2906,13 +2739,7 @@ function extract() {
 
 
   // Keys & frequently used fields
-  const parcelIdFromHtml = getTextOrNull($("#pacel_no"));
-  // CRITICAL: parcel_identifier is REQUIRED and must be a string with minLength: 1
-  // Use parcel_id from seed file, or request_identifier as fallback
-  const parcelId = parcelIdFromHtml ||
-                   (seed && seed.parcel_id) ||
-                   (seed && seed.request_identifier) ||
-                   "UNKNOWN_PARCEL";
+  const parcelId = getTextOrNull($("#pacel_no"));
   // Fallback to seed if HTML element is not found or empty
   const requestIdentifier =(seed &&
       seed.source_http_request &&
@@ -2947,18 +2774,6 @@ function extract() {
     requestIdentifier || parcelId || null,
     source_http_request
   );
-
-  // CRITICAL: Immediate post-extractExtraFeatures cleanup to ensure utility.json is removed
-  // utility.json should NEVER exist - we use utility_N.json files with proper relationships instead
-  try {
-    const utilityJsonPath = path.join(dataDir, "utility.json");
-    if (fs.existsSync(utilityJsonPath)) {
-      fs.unlinkSync(utilityJsonPath);
-      console.log("Post-extractExtraFeatures cleanup: Removed utility.json - using utility_N.json files with relationships instead");
-    }
-  } catch (e) {
-    console.error("Error in post-extractExtraFeatures utility.json cleanup:", e);
-  }
 
   const mandatoryFields={
       source_http_request: source_http_request,
@@ -3059,17 +2874,15 @@ function extract() {
         // console.error("Failed to create new file:", err.message);
       }
     }
-
-    // CRITICAL: Create property_has_lot relationship: parcel (from) → lot (to)
-    // The "from" must be a parcel entity with parcel_identifier (not property)
-    const relationship = {
-      from: { "/": `./parcel.json` },
-      to: { "/": `./lot.json` }
-    };
-    writeJSON(
-      path.join(dataDir, `relationship_property_has_lot.json`),
-      relationship
-    );
+    
+    // const relationship = {
+    //   from: { "/": `./property.json` },
+    //   to: { "/": `./propertyLot.json` }
+    // };
+    // writeJSON(
+    //   path.join("data", `relationship_property_has_lot.json`),
+    //   relationship
+    // );
 
 
   } catch (e) {
@@ -3123,36 +2936,25 @@ function extract() {
   // const buildingType = structuralElementsBuilding1["Building Type"] || null;
 
   // console.log("usecode",useCodeText);
-  let property_type = mapPropertyTypeFromUseCode(propertyUseText || "");
+  const property_type = mapPropertyTypeFromUseCode(propertyUseText || "");
   // console.log("property_type>>",property_type);
   const ownership_estate_type=mapOwnershipEstateTypeFromUseCode(propertyUseText || "");
-  let build_status= mapBuildStatusFromUseCode(propertyUseText || "");
+  const build_status= mapBuildStatusFromUseCode(propertyUseText || "");
   const structure_form = mapStructureFormFromUseCode(propertyUseText || "");
   const property_usage_type = mapPropertyUsageTypeFromUseCode(propertyUseText || "");
 
-  // CRITICAL: property_type is REQUIRED by schema and cannot be null
-  // If mapping fails, default to "LandParcel" as a reasonable fallback (works with VacantLand status)
-  if (!property_type || typeof property_type !== 'string' || property_type.trim() === '') {
-    property_type = "LandParcel";
-  }
 
-  // CRITICAL: Ensure build_status is valid enum value: VacantLand, Improved, or UnderConstruction
-  // If mapping fails or returns null, default to "VacantLand" to match LandParcel property type
-  if (!build_status || typeof build_status !== 'string' || build_status.trim() === '') {
-    build_status = "VacantLand";
-  }
 
-  // Validate that build_status is one of the allowed enum values
-  const validBuildStatuses = ["VacantLand", "Improved", "UnderConstruction"];
-  if (!validBuildStatuses.includes(build_status)) {
-    build_status = "VacantLand";
-  }
-
-  // Validate that property_type is a valid string
-  const validPropertyTypes = ["Cooperative", "Condominium", "Modular", "ManufacturedHousingMultiWide", "Pud", "Timeshare", "2Units", "DetachedCondominium", "Duplex", "SingleFamily", "MultipleFamily", "3Units", "ManufacturedHousing", "ManufacturedHousingSingleWide", "4Units", "Townhouse", "NonWarrantableCondo", "VacantLand", "Retirement", "MiscellaneousResidential", "ResidentialCommonElementsAreas", "MobileHome", "Apartment", "MultiFamilyMoreThan10", "MultiFamilyLessThan10", "LandParcel", "Building", "Unit", "ManufacturedHome"];
-  if (!validPropertyTypes.includes(property_type)) {
-    property_type = "LandParcel";
-  }
+  // For aquaculture/submerged land, default to VacantLand if no type determined
+  // if (!property_type && property_legal_description_text && 
+  //     property_legal_description_text.toLowerCase().includes('aquaculture')) {
+  //   property_type = 'VacantLand';
+  // }
+  
+  // Ensure property_type is never null - default to VacantLand
+  // if (!property_type) {
+  //   property_type = 'VacantLand';
+  // }
 
   // Attempt to extract subdivision from legal description
   let subdivision = null;
@@ -3191,18 +2993,6 @@ function extract() {
   });
   writeJSON(path.join(dataDir, "property.json"), property);
 
-  // CRITICAL: Create parcel.json file - required for property_has_lot relationship
-  // The parcel entity is the "from" in the property_has_lot relationship
-  const parcel = {
-    parcel_identifier: parcelId,
-    request_identifier: requestIdentifier
-  };
-  writeJSON(path.join(dataDir, "parcel.json"), parcel);
-
-  // CRITICAL: Create property_has_lot relationship: parcel (from) → lot (to)
-  // This relationship is created after lot.json is generated (lines 3056)
-  // We'll add the relationship creation right after the lot file is written
-
   // ADDRESS
   try {
     // const siteAddressHtml = $("#site_address").html();
@@ -3218,15 +3008,28 @@ function extract() {
 
     const country_code = "US";
 
-    // CRITICAL: Address schema uses oneOf - provide EITHER unnormalized_address OR normalized fields
-    // Since we have unnormalized_address from input, use that. Do NOT include latitude/longitude (deprecated)
-    // Do NOT mix unnormalized with normalized fields - it's one OR the other
     const address = {
       source_http_request: source_http_request,
       request_identifier: requestIdentifier,
+      // block: block,
+      // city_name: city ? city.toUpperCase() : null,
+      // country_code: country_code,
+      county_name: county_name,
+      latitude: unAddr.latitude ?? null,
+      longitude: unAddr.longitude ?? null,
+      // plus_four_postal_code: null,
+      // postal_code: zip,
+      // range: null,
+      // route_number: null,
+      // section: null,
+      // township: null,
       unnormalized_address: unAddr.full_address
-      // Note: latitude and longitude are deprecated in address schema
-      // Note: county_name should only be included in normalized address form, not with unnormalized_address
+      // state_code: state,
+      // // street_name: street_name ? street_name : null,
+      // // street_post_directional_text: null,
+      // // street_pre_directional_text: null,
+      // // street_number: street_number,
+      // // street_suffix_type: street_suffix_type,
     };
 
     writeJSON(path.join(dataDir, "address.json"), address);
@@ -3235,12 +3038,29 @@ function extract() {
   }
 
 
+  //Mailing Address creation
+  try {
+
+    const mailing_address = {
+      source_http_request: source_http_request,
+      request_identifier: requestIdentifier,
+      latitude:  null,
+      longitude: null,
+      unnormalized_address: mailing_address_text
+    };
+
+    writeJSON(path.join(dataDir, "mailing_address.json"), mailing_address);
+  } catch (e) {
+    console.error("Error extracting mailing address data:", e);
+  }
+
+
   // SALES AND OWNERS CODE BLOCK--------------
   try {
     const salesRows = $("#tblSalesHistory tbody tr");
 
     fs.readdirSync(dataDir).forEach((file) => {
-      if (/^sales_history_\d+\.json$/i.test(file) || /^sales_\d+\.json$/i.test(file) || /^relationship_property_has_sales_history_\d+\.json$/i.test(file)) {
+      if (/^sales_history_\d+\.json$/i.test(file) || /^sales_\d+\.json$/i.test(file)) {
         fs.unlinkSync(path.join(dataDir, file));
       }
     });
@@ -3257,17 +3077,16 @@ function extract() {
       const iso = toISODate(dateTxt);
       const price = parseCurrencyToNumber(priceTxt);
       console.log(granteeRaw,price)
-      // Only create sales_history records when there's a valid price (schema requires purchase_price_amount to be a number, not null)
-      if (iso && price !== null && typeof price === 'number' && Number.isFinite(price)) {
-        const saleObj = {
+      // && price !== null && price > 0
+      if (iso) {
+        sales.push({
           ownership_transfer_date: iso,
           purchase_price_amount: price,
           request_identifier: requestIdentifier, // Include request_identifier for each sale
           source_http_request: source_http_request, // Include the full source_http_request object
           _rawIndex: i, // Internal use for sorting
           grantee_text: granteeRaw,
-        };
-        sales.push(saleObj);
+        });
       }
     });
     console.log("SALES END------",sales)
@@ -3281,27 +3100,10 @@ function extract() {
       const file = path.join(dataDir, saleFileName);
       // Remove _rawIndex and grantee_text before writing to file
       const { _rawIndex, grantee_text, ...saleData } = s;
-
-      // CRITICAL: purchase_price_amount must be a number (not null) per schema
-      // Since it's optional, omit it entirely if not a valid number
-      if (saleData.purchase_price_amount === null ||
-          saleData.purchase_price_amount === undefined ||
-          typeof saleData.purchase_price_amount !== 'number' ||
-          !Number.isFinite(saleData.purchase_price_amount)) {
-        delete saleData.purchase_price_amount;
-      }
-
       writeJSON(file, saleData);
       s._file = `./${saleFileName}`; // Keep _file for relationship linking
       saleFileMap.set(s.ownership_transfer_date, s._file);
       if (!firstSaleFile) firstSaleFile = s._file;
-
-      // Create property→sales_history relationship
-      const relationshipFile = path.join(dataDir, `relationship_property_has_sales_history_${idx + 1}.json`);
-      writeJSON(relationshipFile, {
-        from: { "/": "./property.json" },
-        to: { "/": `./${saleFileName}` }
-      });
     });
 
     const ownerExtraction = extractOwnersFromHtml(
@@ -3479,39 +3281,25 @@ function extract() {
       });
     });
 
-    // Only create mailing_address.json if there are owners to link it to
-    if ((mailingPersonRefs.size > 0 || mailingCompanyRefs.size > 0) && mailing_address_text) {
-      const mailing_address = {
-        source_http_request: source_http_request,
-        request_identifier: requestIdentifier,
-        latitude:  null,
-        longitude: null,
-        unnormalized_address: mailing_address_text
-      };
-      writeJSON(path.join(dataDir, "mailing_address.json"), mailing_address);
-
-      // Create relationships for person → mailing_address
-      mailingPersonRefs.forEach((personRef) => {
-        const relFile = `relationship_person_${personRef
-          .replace("./", "")
-          .replace(".json", "")}_has_mailing_address.json`;
-        writeJSON(path.join(dataDir, relFile), {
-          from: { "/": personRef },
-          to: { "/": "./mailing_address.json" },
-        });
+    mailingPersonRefs.forEach((personRef) => {
+      const relFile = `relationship_person_${personRef
+        .replace("./", "")
+        .replace(".json", "")}_has_mailing_address.json`;
+      writeJSON(path.join(dataDir, relFile), {
+        from: { "/": personRef },
+        to: { "/": "./mailing_address.json" },
       });
+    });
 
-      // Create relationships for company → mailing_address
-      mailingCompanyRefs.forEach((companyRef) => {
-        const relFile = `relationship_company_${companyRef
-          .replace("./", "")
-          .replace(".json", "")}_has_mailing_address.json`;
-        writeJSON(path.join(dataDir, relFile), {
-          from: { "/": companyRef },
-          to: { "/": "./mailing_address.json" },
-        });
+    mailingCompanyRefs.forEach((companyRef) => {
+      const relFile = `relationship_company_${companyRef
+        .replace("./", "")
+        .replace(".json", "")}_has_mailing_address.json`;
+      writeJSON(path.join(dataDir, relFile), {
+        from: { "/": companyRef },
+        to: { "/": "./mailing_address.json" },
       });
-    }
+    });
   } catch (e) {
     console.error("Error extracting sales/owner data:", e);
   }
@@ -3644,7 +3432,7 @@ function extract() {
         to: { "/": `./utility_${utilityIndex}.json` }
       };
       writeJSON(
-        path.join("data", `relationship_layout_to_utility_${utilityIndex}.json`),
+        path.join("data", `relationship_layout_${buildingNumber}_has_utility_${utilityIndex}.json`),
         relationship
       );
     });
@@ -3723,7 +3511,7 @@ function extract() {
               to: { "/": `./layout_${subLayoutIndex}.json` }
             };
             writeJSON(
-              path.join("data", `relationship_layout_${buildingLayoutIndex}_to_layout_${subLayoutIndex}.json`),
+              path.join("data", `relationship_layout_${buildingNumber}_has_layout_${subLayoutIndex}.json`),
               relationship
             );
           }
@@ -3827,7 +3615,7 @@ function extract() {
         to: { "/": `./structure_${structureIndex}.json` }
       };
       writeJSON(
-        path.join("data", `relationship_layout_to_structure_${structureIndex}.json`),
+        path.join("data", `relationship_layout_${buildingNumber}_has_structure_${structureIndex}.json`),
         relationship
       );
     });
@@ -3838,7 +3626,7 @@ function extract() {
 //   try {
 //     const relationshipsToRemove = [
 //       "relationship_property_address.json",
-//       "relationship_property_lot.json",
+//       "relationship_property_lot.json", 
 //       "relationship_property_structure.json",
 //       "relationship_property_utility.json"
 // ];
@@ -3854,124 +3642,7 @@ function extract() {
 //     console.error("Error removing null relationships:", e);
 //   }
 
-  // Final cleanup: Ensure utility.json is removed (it's not used - we use utility_N.json instead)
-  // This file should never exist in the final output. We use utility_N.json files with proper relationships instead.
-  try {
-    // Try multiple path variations to ensure cleanup
-    const pathsToCheck = [
-      path.join(dataDir, "utility.json"),
-      path.resolve(dataDir, "utility.json"),
-      path.join("data", "utility.json"),
-      "./data/utility.json",
-      "data/utility.json"
-    ];
-
-    let removedCount = 0;
-    pathsToCheck.forEach(utilPath => {
-      try {
-        if (fs.existsSync(utilPath)) {
-          fs.unlinkSync(utilPath);
-          removedCount++;
-          console.log(`Removed utility.json at: ${utilPath} - using utility_N.json files with relationships instead`);
-        }
-      } catch (e) {
-        console.error(`Error removing utility.json at ${utilPath}:`, e);
-      }
-    });
-
-    if (removedCount === 0) {
-      console.log("No utility.json file found to remove (expected)");
-    }
-  } catch (e) {
-    console.error("Error during utility.json cleanup:", e);
-  }
-
-  // Additional cleanup: Remove any other unexpected utility-related files
-  try {
-    const unexpectedFiles = ["propertyUtility.json", "property_utility.json", "utility.json"];
-    unexpectedFiles.forEach(filename => {
-      const filepath = path.join(dataDir, filename);
-      if (fs.existsSync(filepath)) {
-        fs.unlinkSync(filepath);
-        console.log(`Removed unexpected file: ${filename}`);
-      }
-    });
-  } catch (e) {
-    console.error("Error cleaning up unexpected files:", e);
-  }
-
-  // Final verification: Absolutely ensure utility.json does not exist
-  try {
-    const finalCheckPath = path.join(dataDir, "utility.json");
-    if (fs.existsSync(finalCheckPath)) {
-      console.error("WARNING: utility.json still exists after cleanup, forcing removal");
-      fs.unlinkSync(finalCheckPath);
-      console.log("Forced removal of utility.json successful");
-    }
-  } catch (e) {
-    console.error("Error in final utility.json verification:", e);
-  }
 
 }
 
-// Run extraction with guaranteed cleanup
-try {
-  extract();
-} finally {
-  // CRITICAL: Post-execution cleanup - absolutely ensure utility.json does not exist
-  // This runs after extract() completes OR if there's an error, to catch any edge cases
-  try {
-    const dataDir = "data";
-    const pathsToClean = [
-      path.join(dataDir, "utility.json"),
-      path.resolve(dataDir, "utility.json"),
-      path.join("data", "utility.json"),
-      "./data/utility.json",
-      "data/utility.json"
-    ];
-
-    let cleanedCount = 0;
-    pathsToClean.forEach(utilPath => {
-      try {
-        if (fs.existsSync(utilPath)) {
-          fs.unlinkSync(utilPath);
-          cleanedCount++;
-          console.log(`Post-execution cleanup: Removed utility.json at ${utilPath} - using utility_N.json files instead`);
-        }
-      } catch (e) {
-        console.error(`Error removing utility.json at ${utilPath}:`, e);
-      }
-    });
-
-    if (cleanedCount === 0) {
-      console.log("Post-execution cleanup: No utility.json found (expected)");
-    }
-  } catch (e) {
-    console.error("Error in post-execution utility.json cleanup:", e);
-  }
-
-  // ABSOLUTE FINAL CHECK: One last verification with fs.readdirSync to catch any edge cases
-  try {
-    const dataDir = "data";
-    if (fs.existsSync(dataDir)) {
-      const files = fs.readdirSync(dataDir);
-      const utilityJsonFiles = files.filter(f => f === "utility.json");
-      if (utilityJsonFiles.length > 0) {
-        console.error("CRITICAL: utility.json still exists after all cleanup attempts!");
-        utilityJsonFiles.forEach(file => {
-          try {
-            const fullPath = path.join(dataDir, file);
-            fs.unlinkSync(fullPath);
-            console.log(`Emergency cleanup: Force deleted ${fullPath}`);
-          } catch (err) {
-            console.error(`Failed to delete ${file} in emergency cleanup:`, err);
-          }
-        });
-      } else {
-        console.log("Final verification: Confirmed utility.json does not exist ✓");
-      }
-    }
-  } catch (e) {
-    console.error("Error in final verification:", e);
-  }
-}
+extract();
